# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/puzzle/create": {
      "post": {
        "tags": [
          "puzzle"
        ],
        "summary": "Create Puzzle",
        "operationId": "create_puzzle_puzzle_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PuzzleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/puzzle/current/{user_id}": {
      "get": {
        "tags": [
          "puzzle"
        ],
        "summary": "Get Current Puzzle",
        "operationId": "get_current_puzzle_puzzle_current__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/puzzle/answer": {
      "post": {
        "tags": [
          "puzzle"
        ],
        "summary": "Submit Answer",
        "operationId": "submit_answer_puzzle_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PuzzleAnswer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PuzzleResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/puzzle/points/{team_id}": {
      "get": {
        "tags": [
          "puzzle"
        ],
        "summary": "Get Team Points",
        "operationId": "get_team_points_puzzle_points__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamPoints"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/puzzle/decay/{team_id}": {
      "post": {
        "tags": [
          "puzzle"
        ],
        "summary": "Decay Points",
        "operationId": "decay_points_puzzle_decay__team_id__post",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PuzzleAnswer": {
        "properties": {
          "puzzle_id": {
            "type": "integer",
            "title": "Puzzle Id",
            "description": "ID of the puzzle being answered"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Player's answer to the puzzle"
          }
        },
        "type": "object",
        "required": [
          "puzzle_id",
          "answer"
        ],
        "title": "PuzzleAnswer",
        "description": "Puzzle Answer: Request to submit a puzzle answer"
      },
      "PuzzleCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "memory",
              "spatial",
              "concentration",
              "multitasking"
            ],
            "title": "Type",
            "description": "Type of puzzle to create"
          },
          "game_session_id": {
            "type": "integer",
            "title": "Game Session Id",
            "description": "ID of the game session"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "ID of the user for this puzzle"
          }
        },
        "type": "object",
        "required": [
          "type",
          "game_session_id",
          "user_id"
        ],
        "title": "PuzzleCreate",
        "description": "Puzzle Create: Request to create a new puzzle"
      },
      "PuzzleResult": {
        "properties": {
          "correct": {
            "type": "boolean",
            "title": "Correct",
            "description": "Whether the answer was correct"
          },
          "awarded_to_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awarded To User Id",
            "description": "ID of user who received points"
          },
          "points_awarded": {
            "type": "integer",
            "title": "Points Awarded",
            "description": "Number of points awarded"
          },
          "next_puzzle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Puzzle Id",
            "description": "ID of next puzzle (if any)"
          },
          "next_puzzle": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Next Puzzle",
            "description": "Next puzzle data (if any)"
          }
        },
        "type": "object",
        "required": [
          "correct",
          "points_awarded"
        ],
        "title": "PuzzleResult",
        "description": "Puzzle Result: Puzzle result response model"
      },
      "PuzzleState": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Puzzle ID"
          },
          "type": {
            "type": "string",
            "enum": [
              "memory",
              "spatial",
              "concentration",
              "multitasking"
            ],
            "title": "Type",
            "description": "Puzzle type"
          },
          "data": {
            "title": "Data",
            "description": "Puzzle-specific data"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "failed"
            ],
            "title": "Status",
            "description": "Puzzle status"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer",
            "description": "Correct answer for the puzzle"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "data",
          "status",
          "correct_answer"
        ],
        "title": "PuzzleState",
        "description": "Puzzle State: Puzzle state response model"
      },
...
      "TeamPoints": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id",
            "description": "Team ID"
          },
          "players": {
            "items": {},
            "type": "array",
            "title": "Players",
            "description": "List of players with their points"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "players"
        ],
        "title": "TeamPoints",
        "description": "Team Points: Team points response model"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}